Switch complex conjugated element in inner product
This change should lead to less error-prone code.
Since for approximation purposes it is a common notation
that the first argument of the inner product is time
dependent and the time dependent part
will be excluded from the inner product. If we
put complex conjugation on the first argument one
have to consider the complex conjugated time dependent
part in subsequent code. Of course the argument order could
be switched by each function call but it remains: some consideration is necessary and some of these calls are
hard coded inside pyinduct functions.